+++ server/src/entities/Users.ts
@@ -17,8 +17,14 @@ export class Users {
   @Column("character varying", { name: "username", length: 55 })
   username: string;
 
-  @Column("character varying", { name: "password", length: 55 })
+  @Column("character varying", { name: "password", length: 60 })
   password: string;
+
+  @Column("character varying", { name: "email", length: 55 })
+  email: string;
+
+  @Column("bytea", { name: "avatar", nullable: true })
+  avatar: Buffer;
 
   @OneToMany(() => Books, (books) => books.user)
   books: Books[];