+++ server/src/routes/v1/login/login.ts
@@ -3,6 +3,7 @@ import { AppDataSource } from "../../../data-source";
 import { Users } from "../../../entities/Users";
 import { sign } from "jsonwebtoken";
 import { JWT_SECRET } from "../../../env";
+import { crypter } from "../../../middlewares/crypter";
 
 export const login = async (req: Request, res: Response) => {
   // Get the username and password from the request body
@@ -14,12 +15,27 @@ export const login = async (req: Request, res: Response) => {
   try {
     // Validate the username and password against your data source (for example, a database)
     const repo = AppDataSource.getRepository(Users);
-    const login = await repo.findBy({
+    const loginByUsername = await repo.findBy({
       username: username,
-      password: password,
     });
 
-    if (login.length > 0) {
+    const loginByEmail = await repo.findBy({
+      email: username,
+    });
+
+    if (loginByUsername.length > 0 || loginByEmail.length > 0) {
+      const login = loginByUsername.length > 0 ? loginByUsername : loginByEmail;
+      console.log(login);
+
+      const goodPassword = await crypter.compare(password, login[0].password);
+
+      if (!goodPassword) {
+        res.status(401).json({
+          message: "Invalid password !",
+        });
+        return;
+      }
+
       const ONE_MONTH = 1000 * 60 * 60 * 24 * 30;
       const token = sign({ username: username }, JWT_SECRET, {
         expiresIn: ONE_MONTH,
@@ -32,7 +48,7 @@ export const login = async (req: Request, res: Response) => {
       });
     } else {
       res.status(401).json({
-        message: "Invalid username or password",
+        message: "Invalid username !",
       });
     }
   } catch (error) {